+++ src/main/java/club/lemos/common/node/TreeNode.java
@@ -14,9 +14,9 @@ public class TreeNode extends BaseNode implements Serializable {
 
 	private String title;
 
-	private Integer key;
+	private Long key;
 
-	private Integer value;
+	private Long value;
 
 	private Integer type;
 }
